IGNITE-11867 Force clearing of newly owned evicted/rented partition.

+++ modules/core/src/main/java/org/apache/ignite/internal/processors/cache/persistence/file/FilePageStore.java
@@ -276,7 +276,7 @@ public class FilePageStore implements PageStore {
                 if (fileIO != null) // Ensure the file is closed even if not initialized yet.
                     fileIO.close();
 
-                if (delete)
+                if (delete && cfgFile.exists())
                     Files.delete(cfgFile.toPath());
 
                 return;
